feat: add authcontainer
