+++ my-app/src/App.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Header from './components/header/Header';
-import AuthFormLogin from './components/authForm/AuthFormLogin';
+import AuthFormContainer from './components/authForm/AuthFormContainer';
 import ErrorPage from './components/errorPage/ErrorPage';
 import ScrollButton from './components/common/scrollButton/ScrollButton';
-import AuthFormRegister from './components/authForm/AuthFormRegister';
 import Footer from './components/common/footer/Footer';
 import HomePage from './components/pages/homePage';
 import DetailsPage from './components/pages/detailsPage';
@@ -20,8 +19,8 @@ const App: React.FC = () => {
     <BrowserRouter basename='/React-project'>
       <Header />
       <Routes>
-        <Route path='/register' element={<AuthFormRegister />} />
-        <Route path='/auth' element={<AuthFormLogin />} />
+        <Route path='/register' element={<AuthFormContainer isRegisterForm={true} />} />
+        <Route path='/auth' element={<AuthFormContainer isRegisterForm={false} />} />
         <Route path='/' element={<HomePage />} />
         <Route path='*' element={<ErrorPage />} />
         <Route path='/movie/:id' element={<DetailsPage />} />
